+++ spec/services/random_lottery_spec.rb
@@ -108,7 +108,8 @@ RSpec.describe RandomLottery do
   end
 
   def valid_mock_draw
-    instance_spy(Draw, lottery_assignments: [], lottery?: true).tap do |d|
+    # persisting data because of the lottery notification e-mail
+    create(:draw_in_lottery).tap do |d|
       allow(ObjectsForLotteryQuery).to \
         receive(:call).with(draw: d).and_return([])
     end